Optimize removeRange by removing the loop
The loop only has to be done when looping over all. At that point we should do the cleanup, check
whether any range is still referenced.

By saving the ranges in a weakset as well, we can detect manual removals as well as automatic ones
because the weakref is cleaned up.